fix: user controller web interface and status codes
Previously a successful user registration returned a 200, it now returns a 201. This bug revealed a larger hole in the testing of the UserController. Integration tests have been added to test the response codes resulting from this controller.

Previously validating an email address was conducted using a POST request when it should have been a GET request.